backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
pyodbc
python-multipart
aioodbc
